+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.4] - 2025-12-1
+
+### Improvements
+- Improved the speed and effectiveness of computing the image histograms on realistic data
+
+### Bug Fixes
+- Fixed bug where progress dialog would close when loading imagery before the histogram creationg progress dialog would open 
+- Forced loaded imagery to cast to float32. All image processing algorithms assume data are floating point values.
+
 ## [1.6.3] - 2025-11-30
 
 ### Improvements
@@ -135,6 +144,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed bug where tooltips did not take into account imagery row / column offsets.
 - Fixed bug where imagery produced by algorithms did not have pre-computed histograms (which improves playback performance)
 
+[1.6.4]: https://github.com/hartzell-stephen-me/vista/releases/tag/1.6.4
 [1.6.3]: https://github.com/hartzell-stephen-me/vista/releases/tag/1.6.3
 [1.6.2]: https://github.com/hartzell-stephen-me/vista/releases/tag/1.6.2
 [1.6.1]: https://github.com/hartzell-stephen-me/vista/releases/tag/1.6.1